Simplify appending to env var JAVACMD_OPTIONS
In default Bash, it is not an error to refer to an undeclared (or empty)
variable. In that case, it will just expand to an empty string.

So there's no need to do the prior check, you can just append the new value
right away.

Also, in Bash you can safely set and export on the same command.